+++ packages/react-ui/src/theme/carto-theme.js
@@ -1,4 +1,7 @@
-import { themeComponentsOverrides, themeComponentsProps } from './sections/components';
+import { dataDisplayOverrides } from './sections/components/dataDisplay';
+import { buttonsOverrides } from './sections/components/buttons';
+import { formsOverrides } from './sections/components/forms';
+import { navigationOverrides } from './sections/components/navigation';
 import { CssBaseline } from './sections/cssBaseline';
 import { commonPalette } from './sections/palette';
 import { themeShadows } from './sections/shadows';
@@ -99,13 +102,17 @@ export const cartoThemeOptions = {
     snackbar: 1400,
     tooltip: 1500
   },
-  overrides: {
-    MuiCssBaseline: {
-      ...CssBaseline
-    },
-    ...themeComponentsOverrides
-  },
 
-  // Props
-  props: { ...themeComponentsProps }
+  // Styles and props overrides for components
+  components: {
+    MuiCssBaseline: {
+      styleOverrides: {
+        ...CssBaseline
+      }
+    },
+    ...buttonsOverrides,
+    ...formsOverrides,
+    ...navigationOverrides,
+    ...dataDisplayOverrides
+  }
 };